Add linter rule to require React
