+++ components/containerDetailsButton/styles.jsx
