+++ pages/500.jsx
@@ -19,6 +19,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
+import React from "react";
 import Link from "next/link";
 
 export default function Custom404() {
