+++ pages/contacts/add.jsx
@@ -19,6 +19,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
+import React from "react";
 import AddContactPage from "../../components/addContact";
 
 export default function Contacts() {
